no more error in backend, one last error to update password

+++ back/controllers/img.js
@@ -71,14 +71,11 @@ exports.getImg = (req, res) => {
 };
 
 exports.updateImg = (req, res) => {
-    const id = req.params.id;
     const token = req.headers.authorization.split(' ')[1];
     const decodedToken = jwt.verify(token, process.env.JWT_SECRET_KEY);
     const userId = decodedToken.user_id;
-
-    if (id === userId) {
         Img.findOne({
-            where: { id: id }
+            where: { id: req.params.id }
         }).then(
             (image) => {
                 if (req.file) {
@@ -97,6 +94,7 @@ exports.updateImg = (req, res) => {
                     }
                 }).then(
                     (image) => {
+                        console.log('image updated');
                         res.status(200).json(image)
                     }
                 ).catch(
@@ -106,22 +104,49 @@ exports.updateImg = (req, res) => {
         ).catch(
             error => res.status(404).json({ error })
         );
-    } 
 };
 
-exports.deleteImg = async (req, res) => {
-    await Img.findOne({
-        where: {
-            id: req.params.id
+// exports.deleteImg = (req, res) => {
+//     Img.findOne({
+//         where: {
+//             id: req.params.id
+//         }
+//     })
+//     .then(img => {
+//         const filename = img.image_url.split('/images/')[1];
+//         fs.unlink(`images/${filename}`, () => {
+//             Img.destroy({
+//                 where: {id: req.params.id}
+//             })
+//         });
+//     })
+//     .catch(error => res.status(400).json({error}));
+// };
+
+exports.deleteImg = (req, res) => {
+    Img.findOne({
+        where: { id: req.params.id }
+    }).then(
+        (image) => {
+            console.log('req.params.id:' + req.params.id);
+            if (image.image_url) {
+                const filename = image.image_url.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Img.destroy({
+                        where: { id: req.params.id }
+                    }).then(() => res.status(200).json({ message: 'Image deleted'}))
+                    .catch((error) => res.status(400).json({error}))
+                });
+            }
+            // Img.destroy({
+            //     where: { id: req.params.id }
+            // }).then(() => res.status(200).json({ message: 'Image deleted'}))
+            // .catch((error) => res.status(400).json({error}))
         }
-    })
-    .then(img => {
-        const filename = img.image_url.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            Img.destroy({
-                where: {id: req.params.id}
-            })
-        });
-    })
-    .catch(error => res.status(400).json({error}));
-};
+    ).catch(
+        (error) => {
+            console.error(error);
+            res.status(500).json({error: error})
+        }
+    );
+}
